6-size.c

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int main(void)
 	long long int d;
 	float f;
 
-printf("Size of a char: %zu byte(s)\n",sizeof(a));
-printf("Size of an int: %zu byte(s)\n",sizeof(b));
-printf("Size of a long int: %zu byte(s)\n",sizeof(c));
-printf("size of a long long int: %zu byte(s)\n",sizeof(d));
-printf("size of a float: %zu byte(s)\n",sizeof(f));
+printf("Size of a char: %zu byte(s)\n", (unsigned long)sizeof(a));
+printf("Size of an int: %zu byte(s)\n", (unsigned long)sizeof(b));
+printf("Size of a long int: %zu byte(s)\n",(unsigned long)sizeof(c));
+printf("size of a long long int: %zu byte(s)\n", (unsigned long)sizeof(d));
+printf("size of a float: %zu byte(s)\n", (unsigned long)sizeof(f));
 return (0);
 }
